# service-resource-manager-implementation/src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: SRM Controller API
  description: |
    API exposed by SRM for "PaaS" - based interaction with NFV MANO.
  termsOfService: http://swagger.io/terms/
  contact:
    email: dlaskaratos@intracom-telecom.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
...
servers:
- url: http://vitrualserver:8080/piedge-connector/2.0.0
paths:
...
  /copy-artefact:
    post:
      tags:
      - Artefact Management
      summary: Copies artefact from source repository to destination repository
      operationId: copy_artifact
      requestBody:
        description: Artifact details including image name, tag, source repository username and password tec
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyArtifactModel'
      responses:
        "200":
          description: Artifact successfully copied
        "400":
          description: Mandatory fields missing
      x-openapi-router-controller: src.controllers.artifact_controller
  /artefact-exists:
    post:
      tags:
      - Artefact Management
      summary: Check if artefact exists in given repository
      operationId: artifact_exists
      requestBody:
        description: Artefact details including image name, tag, source repository username and password tec
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactExistsModel'
      responses:
        "200":
          description: Artifact exists
        "404":
          description: Artifact does not exist
        "400":
          description: Mandatory fields missing        
      x-openapi-router-controller: src.controllers.artifact_controller   
  /serviceFunction:
    post:
      tags:
      - Service Functions Catalogue
      summary: Register Service.
      # security:
      #   - jwt: [ ]
      operationId: register_service_function
      requestBody:
        description: Registration method to save service function into database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceFunctionRegistrationRequest'
      responses:
        "200":
          description: Service function registered
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
    get:
      tags:
      - Service Functions Catalogue
      summary: Returns service functions from the catalogue.
      operationId: get_service_functions
      # security:
      #   - jwt: [ ]
      responses:
        "200":
          description: Returns service functions from the catalogue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/appsResponse'
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
  /serviceFunction/{serviceFunctionId}:
    delete:
      tags:
      - Service Functions Catalogue
      summary: Deregister service.
      # security:
      #   - jwt: [ ]
      operationId: deregister_service_function
      parameters:
      - name: serviceFunctionId
        in: path
        description: Returns a  specific service function from the catalogue.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Service function unregistered
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
    get:
      tags:
      - Service Functions Catalogue
      summary: Returns a specific service function from the catalogue.
      # security:
      #   - jwt: [ ]
      operationId: get_service_function
      parameters:
      - name: serviceFunctionId
        in: path
        description: Returns a  specific service function from the catalogue.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Returns the service function info status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/appsResponse_apps'
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
  /node:
    get:
      tags:
      - Nodes
      summary: Returns the edge nodes status.
      # security:
      #   - jwt: [ ]
      operationId: get_nodes
      responses:
        "200":
          description: Returns the edge nodes status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/nodesResponse'
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
  /deployedServiceFunction:
    post:
      tags:
      - Service Functions Instances
      summary: Request to deploy a Service function (from the catalogue) to an edge
        node.
      operationId: deploy_service_function
      # security:
      #   - jwt: [ ]
      requestBody:
        description: Deploy Service Function.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployApp'
      responses:
        "200":
          description: App deployed.
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
    get:
      tags:
      - Service Functions Instances
      summary: Returns deployed apps.
      operationId: get_deployed_service_functions
      # security:
      #   - jwt: [ ]
      responses:
        "200":
          description: Returns service functions from the catalogue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/appsResponse'
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
  /deployedServiceFunction/{app_id}:
    delete:
      tags:
      - Service Functions Instances
      summary: Deletes a deployed Service function.
      operationId: delete_deployed_service_function
      # security:
      #   - jwt: [ ]
      parameters:
      - name: app_id
        in: path
        description: Represents a service function from the running deployments.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Deployed service function  deleted
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller
    get:
      tags:
      - Service Functions Instances
      summary: Returns deployed apps.
      operationId: get_deployed_service_function
      parameters:
      - name: app_id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Returns service functions from the catalogue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/appsResponse'
        "405":
          description: Method not allowed
      x-openapi-router-controller: src.controllers.edge_cloud_management_controller  
...
components:
  schemas:
...
    ServiceFunctionRegistrationRequest:
      type: object
      properties:
        service_function_name:
          type: string
          example: Kibana
        service_function_image:
          type: string
          example: kibana:7.15.2
        service_function_type:
          type: string
          example: Container
        application_ports:
          type: array
          example:
          - 5601
          items:
            type: integer
        autoscaling_policies:
          type: array
          items:
            $ref: '#/components/schemas/autoscaling_policy'
        required_volumes:
          type: array
          items:
            $ref: '#/components/schemas/volume'
        required_env_parameters:
          type: array
          items:
            $ref: '#/components/schemas/env_parameter_name'
        privileged:
          type: boolean
          example: false
...
    autoscaling_policy_metric:
      type: object
      properties:
        metric:
          example: cpu
          type: string
        limit:
          type: string
          example: 1000m
        request:
          type: string
          example: 600m
        util_percent:
          type: number
          example: 60
        is_default:
          type: boolean
          example: true
    autoscaling_policy:
      type: object
      properties:
        policy:
          type: string
          example: maximize-performance
        monitoring_metrics:
          type: array
          items:
            $ref: '#/components/schemas/autoscaling_policy_metric'
...
    volume:
      type: object
      properties:
        name:
          type: string
          example: volumeconfig
        path:
          type: string
          example: /data/config
        hostpath:
          type: string
          example: /data/config
    env_parameter_name:
      type: object
      properties:
        name:
          type: string
      example:
        name: ELASTICSEARCH_URL
...
    nodesResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: openness
        location:
          type: string
        serial:
          type: string
        node_type:
          type: string
      example:
        serial: 146.124.106.179
        name: compute1
        location: Peania_19002_Athens
        id: 237d11c4-aca6-4845-9538-ba7b3e89c0b6
        node_type: server
    appsResponse:
      type: object
      properties:
        apps:
          type: array
          items:
            $ref: '#/components/schemas/appsResponse_apps'
      example:
        apps:
        - id: id
        - id: id
...
    appsResponse_apps:
      type: object
      properties:
        id:
          type: string
      example:
        id: id
...
    CopyArtifactModel:
      type: object
      properties:
        src_registry:
          type: string
          example: http://dockerhub.io
        src_image_name:
          type: string
          example: mongodb
        src_image_tag:
          type: string
          example: latest
        dst_registry:
          type: string
          example: http://dockerhub.io
        dst_image_name:
          type: string
          example: mongodb
        dst_image_tag:
          type: string
          example: latest
        src_username:
          type: string
          example: user123
        src_password:
          type: string
          example: 1234
        dst_username:
          type: string
          example: user123 
        dst_password:
          type: string
          example: 1234
    ArtifactExistsModel:
      type: object
      properties:
        registry_url:
          type: string
          example: http://dockerhub.io
        artefact_name:
          type: string
          example: mongodb
        artefact_tag:
          type: string
          example: latest
        username:
          type: string
          example: user123
        password:
          type: string
          example: 1234
    DeployApp:
      type: object
      properties:
        appId:
          type: string
          example: "123456789"
        appZones:
          type: array
          items:
            $ref: '#/components/schemas/EdgeCloudZone'
    EdgeCloudZone:
      type: object
      properties:
        edgeCloudZoneId:
          type: string
          example: "123456789"
        edgeCloudZoneName:
          type: string
          example: Test
        edgeCloudZoneStatus:
          type: string
          example: active
        edgeCloudProvider:
          type: string
          example: Test
        edgeCloudRegion:
          type: string
          example: Test        
  securitySchemes:
    registry_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://serviceregistry.swagger.io/oauth/dialog
          scopes:
            write:services: modify services in your account
            read:services: read your services
      x-tokenInfoFunc: src.controllers.authorization_controller.check_registry_auth
      x-scopeValidateFunc: src.controllers.authorization_controller.validate_scope_registry_auth
    api_key:
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: src.controllers.authorization_controller.check_api_key
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: src.controllers.authorization_controller.decode_token